+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PVPCPlanner"
+        android:name=".AppClass"
         tools:targetApi="31">
         <activity
             android:name=".ui.Main.MainActivity"
