+++ app/src/main/java/com/codingpit/pvpcplanner/AppClass.kt
@@ -1,0 +1,8 @@
+package com.codingpit.pvpcplanner
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class AppClass : Application() {
+}
