+++ app/src/main/java/com/codingpit/pvpcplanner/data/Repository.kt
@@ -1,23 +1,7 @@
 package com.codingpit.pvpcplanner.data
 
-import android.util.Log
-import com.codingpit.pvpcplanner.data.remote.RemoteDataSource
 import com.codingpit.pvpcplanner.domains.models.PVPCModel
-import javax.inject.Inject
 
-class Repository @Inject constructor (
-    private val remoteDataSource: RemoteDataSource,
-) {
-    suspend fun getPrices(): List<PVPCModel> {
-        try {
-            Log.d("Repository", "Fetching data...")
-            val responseRemote = remoteDataSource.getPrices()
-            Log.d("Repository", "data fetched successfully")
-            Log.w("Repository", responseRemote.toString())
-            return responseRemote
-        } catch (e: Exception) {
-            Log.e("Repository", "Error fetching data", e)
-            return emptyList() // Or handle the error as needed
-        }
-    }
+interface Repository{
+    suspend fun getPrices(): List<PVPCModel>
 }
