+++ app/src/main/java/com/codingpit/pvpcplanner/di/NetworkModule.kt
@@ -1,6 +1,8 @@
 package com.codingpit.pvpcplanner.di
 
 import com.codingpit.pvpcplanner.data.remote.PVPCApi
+import com.codingpit.pvpcplanner.data.remote.RemoteDataSource
+import com.codingpit.pvpcplanner.data.remote.RemoteDataSourceImpl
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
@@ -42,5 +44,12 @@ object NetworkModule {
             .build()
 
     @Provides
-    fun provideBantuApi(retrofit: Retrofit): PVPCApi = retrofit.create(PVPCApi::class.java)
+    fun providePVPCApi(retrofit: Retrofit): PVPCApi = retrofit.create(PVPCApi::class.java)
+
+    @Provides
+    fun provideRemoteDataSource(remoteDataSourceImpl: RemoteDataSourceImpl): RemoteDataSource {
+        return remoteDataSourceImpl
+    }
 }
+
+
