+++ app/src/main/java/com/codingpit/pvpcplanner/di/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.codingpit.pvpcplanner.di
+
+import com.codingpit.pvpcplanner.data.Repository
+import com.codingpit.pvpcplanner.data.RepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Binds
+    abstract fun bindRepository(repositoryImpl: RepositoryImpl): Repository
+
+}
